-fix: ??????????? ?? ????????? ? ??????, ????? ???????? ?????? ????????? ?? ?????? ??????, ??????? ??????????? ?? ?????????? ????? [no_wornings]

+++ src/scintilla/win32/PlatWin.cxx
@@ -2108,7 +2108,10 @@ double ElapsedTime::Duration(bool reset) {
 		LARGE_INTEGER lBegin;
 		lBegin.HighPart = bigBit;
 		lBegin.LowPart = littleBit;
-		double elapsed = lEnd.QuadPart - lBegin.QuadPart;
+//!-start-[no_wornings]
+//		double elapsed = lEnd.QuadPart - lBegin.QuadPart;
+		double elapsed = static_cast<double>(lEnd.QuadPart - lBegin.QuadPart);
+//!-end-[no_wornings]
 		result = elapsed / static_cast<double>(frequency.QuadPart);
 	} else {
 		endBigBit = clock();
