+++ src/scite/win32/SciTEWinBar.cxx
@@ -588,7 +588,7 @@ void SciTEWin::SetToolBar() {
 		tbb[i].idCommand = barbuttons[i].cmd;
 		tbb[i].iBitmap = barbuttons[i].id;
 		tbb[i].fsState = TBSTATE_ENABLED;
-		tbb[i].fsStyle = -1 == tbb[i].iBitmap ? TBSTYLE_SEP : TBSTYLE_BUTTON;
+		tbb[i].fsStyle = static_cast<BYTE>(-1 == tbb[i].iBitmap ? TBSTYLE_SEP : TBSTYLE_BUTTON);
 		tbb[i].dwData = 0;
 		tbb[i].iString = 0;
 	}
